song menu extracted
