+++ song.rb
@@ -12,5 +12,4 @@ class Song
     @title = title
     @premiere = premiere
   end
-
 end
